explode utils file into separated functions to improve the organization and to remove the circular dependencies.
