+++ react-components/src/components/RuleBasedOutputs/core/createRuleStyling.ts
@@ -1,0 +1,5 @@
+import { NodeAppearance } from "@cognite/reveal";
+
+export function createRuleStyling(appearance: NodeAppearance): NodeAppearance {
+  return { renderGhosted: false, visible: true, ...appearance };
+}
