+++ react-components/src/components/RuleBasedOutputs/core/getFdmPropertyTrigger.ts
@@ -1,0 +1,17 @@
+import { FdmPropertyType } from "../../Reveal3DResources/types";
+import { FdmRuleTrigger } from "../types";
+
+export function getFdmPropertyTrigger<T>(
+  fdmPropertyTrigger: FdmPropertyType<unknown> | undefined,
+  trigger: FdmRuleTrigger
+): T | undefined {
+  if (fdmPropertyTrigger === undefined) return;
+
+  const space = fdmPropertyTrigger[trigger.key.space];
+  const instanceProperties = space?.[
+    `${trigger.key.view.externalId}/${trigger.key.view.version}`
+  ] as FdmPropertyType<unknown>;
+  const property = instanceProperties?.[trigger.key.property] as T;
+
+  return property;
+}
