+++ react-components/src/components/RuleBasedOutputs/core/getRuleTriggerTypes.ts
@@ -1,0 +1,7 @@
+import { RuleWithOutputs, TriggerType } from "../types";
+import { getExpressionTriggerTypes } from "./getExpressionTriggerTypes";
+
+export function getRuleTriggerTypes(ruleWithOutput: RuleWithOutputs): TriggerType[] | undefined {
+  if (ruleWithOutput.rule.expression === undefined) return;
+  return getExpressionTriggerTypes(ruleWithOutput.rule.expression);
+}
