+++ react-components/src/components/RuleBasedOutputs/core/getTimeseriesExternalIdFromNumericExpression.ts
@@ -1,0 +1,14 @@
+import { isFdmTrigger, isMetadataTrigger } from "../typeGuards";
+import { NumericExpression } from "../types";
+
+export const getTimeseriesExternalIdFromNumericExpression = (
+  expression: NumericExpression
+): string[] | undefined => {
+  const trigger = expression.trigger;
+
+  if (isMetadataTrigger(trigger)) return;
+
+  if (isFdmTrigger(trigger)) return;
+
+  return [trigger.externalId];
+};
